Plantform_Backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plantform-backend"
version = "0.1.0"
description = "Plantform backend scaffold with lightweight API modules for auth, nodes, files, and trade flows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14,<2.0",
    "fastapi>=0.115,<1.0",
    "httpx>=0.28,<1.0",
    "psycopg[binary]>=3.2,<4.0",
    "pydantic-settings>=2.7,<3.0",
    "sqlalchemy>=2.0.36,<3.0",
    "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
    "alembic>=1.14,<2.0",
    "httpx>=0.28,<1.0",
    "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["backend_app*"]
